moved firmness scale to seperate component
